// Project-Web/src/app/openapi-api-models/model/standardUserDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StandardUserDto {
    id?: number;
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    city: string;
    county: string;
}